+++ Framework/Model/Model.h
@@ -20,22 +20,22 @@ public:
 	Model();
 	~Model();
 
-	UINT BoneCount() { return bones.size(); }
+	UINT BoneCount() { return static_cast<UINT>(bones.size()); }
 	vector<ModelBone *>& Bones() { return bones; }
 	ModelBone* BoneByIndex(UINT index) { return bones[index]; }
 	ModelBone* BoneByName(wstring name);
 
-	UINT MeshCount() { return meshes.size(); }
+	UINT MeshCount() { return  static_cast<UINT>(meshes.size()); }
 	vector<ModelMesh *>& Meshes() { return meshes; }
 	ModelMesh* MeshByIndex(UINT index) { return meshes[index]; }
 	ModelMesh* MeshByName(wstring name);
 
-	UINT MaterialCount() { return materials.size(); }
+	UINT MaterialCount() { return static_cast<UINT>(materials.size()); }
 	vector<Material *>& Materials() { return materials; }
 	Material* MaterialByIndex(UINT index) { return materials[index]; }
 	Material* MaterialByName(wstring name);
 
-	UINT ClipCount() { return clips.size(); }
+	UINT ClipCount() { return static_cast<UINT>(clips.size()); }
 	vector<ModelClip *>& Clips() { return clips; }
 	ModelClip* ClipByIndex(UINT index) { return clips[index]; }
 	ModelClip* ClipByName(wstring name);
