+++ Framework/Renders/Shader.h
@@ -20,7 +20,7 @@ public:
 
 	void Dispatch(UINT technique, UINT pass, UINT x, UINT y, UINT z);
 
-	UINT PassCount(UINT techIndex = 0) { return techniques[techIndex].Passes.size(); }
+	UINT PassCount(UINT techIndex = 0) { return static_cast<UINT>( techniques[techIndex].Passes.size() ); }
 
 
 	ID3DX11EffectVariable* Variable(string name);
